Added part rotation. Gave extra parts a solidity attribute.
Extra part solidity is represented with slightly darker grid cursor squares. Also added a couple of extra tutorial dialogs; one showing rotation, and one reminding the player that the controls are on the itch page. Lastly, added a couple of details to the main menu. I definitely made the right move not trying to implement rotation during the jam it was pretty time consuming figuring out how to get all the rotated parts to draw correctly in both the construction screen as well as the driving section. Got there in the end though!